requirements.txt
streamlit
tweepy
transformers
torch
pandas
matplotlib
altair
python-dotenv